requirements.txt
google-cloud-storage
fastapi
pyyaml
setuptools
wheel


